Convert old sessions file to ts; remove vm stuff
